Rename translate to translator
