+++ internal/translator/testdata/basic_ingress/input.yaml
