+++ internal/translator/testdata/basic_ingress/output.yaml
