+++ internal/translator/testdata/health-checks/input.yaml
