+++ internal/translator/testdata/health-checks/output.yaml
