+++ internal/translator/testdata/lb-diff-strategy/input.yaml
