+++ internal/translator/testdata/lb-diff-strategy/output.yaml
