+++ internal/translator/testdata/lb-strategy/input.yaml
