+++ internal/translator/testdata/lb-strategy/output.yaml
