+++ internal/translator/testdata/multiple-upstreams/input.yaml
