+++ internal/translator/testdata/multiple-upstreams/output.yaml
