+++ internal/translator/testdata/request-timeout/input.yaml
