+++ internal/translator/testdata/request-timeout/output.yaml
