+++ internal/translator/testdata/tls_ingressroute/output.yaml
