+++ internal/translator/testdata/weight-shifting/input.yaml
