+++ internal/translator/testdata/weight-shifting/output.yaml
